reflection-agent/pyproject.toml
[project]
name = "reflection-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "isort>=6.1.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.34",
    "langgraph>=0.6.8",
    "python-dotenv>=1.1.1",
]
